Time: 31 ms (89.00%), Space: 16.4 MB (20.13%) - LeetHub

+++ 871-minimum-number-of-refueling-stops/871-minimum-number-of-refueling-stops.cpp
@@ -6,7 +6,7 @@ public:
         cout.tie(0);
         
         priority_queue<int> q;
-        long f = startFuel, ans = 0;
+        int f = startFuel, ans = 0;
         for (int i = 0; i < s.size(); ++i) {
              if (f >= s[i][0]) q.push(s[i][1]);
              else {
